requirements.txt
werkzeug
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
Cython==0.29.6
falcon==1.4.1
idna==2.8
IPS-Common-Library==0.1.15
IPS-Common-Library-DB==0.1.5
more-itertools==6.0.0
numexpr==2.6.9
numpy==1.16.2
pandas==0.24.2
pbr==5.1.3
pluggy==0.9.0
py==1.8.0
pycparser==2.19
PyMySQL==0.9.3
pytest==4.3.1
python-dateutil==2.8.0
python-mimeparse==1.6.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.1
urllib3==1.24.1
waitress==1.2.1
falcon-multipart==0.2.0